Add new linelum_kernels.py module
