+++ diffsky/experimental/kernels/tests/test_linelum_kernels.py
@@ -1,0 +1,11 @@
+""""""
+
+from .. import linelum_kernels
+
+
+def test_mc_specphot_kern():
+    pass
+
+
+def test_specphot_kern():
+    pass
